Fixes #177: Allow [site] and [site].[env] to be used rather than appUuid or EnvUuid.
