+++ src/Command/Api/ApiCommandBase.php
@@ -6,6 +6,14 @@ use Acquia\Cli\Command\CommandBase;
 use AcquiaCloudApi\Exception\ApiErrorException;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Output\OutputInterface;
+use Symfony\Component\Validator\Constraints\Length;
+use Symfony\Component\Validator\Constraints\NotBlank;
+use Symfony\Component\Validator\Constraints\Regex;
+use Symfony\Component\Validator\Constraints\Uuid;
+use Symfony\Component\Validator\Constraints\UuidValidator;
+use Symfony\Component\Validator\Constraints\Valid;
+use Symfony\Component\Validator\Exception\ValidatorException;
+use Symfony\Component\Validator\Validation;
 
 /**
  * Class ApiCommandBase.
@@ -53,14 +61,9 @@ class ApiCommandBase extends CommandBase {
    */
   protected function initialize(InputInterface $input, OutputInterface $output) {
     parent::initialize($input, $output);
-
-    if ($input->hasArgument('applicationUuid') && !$input->getArgument('applicationUuid')) {
-      $output->writeln('Inferring Cloud Application UUID for this command since none was provided...');
-      if ($application_uuid = $this->determineCloudApplication()) {
-        $output->writeln("Set application uuid to <options=bold>$application_uuid</>");
-        $input->setArgument('applicationUuid', $application_uuid);
-      }
-    }
+    $this->convertApplicationAliastoUuid($input);
+    $this->fillMissingApplicationUuid($input, $output);
+    $this->convertEnvironmentAliasToUuid($input);
   }
 
   /**
@@ -204,6 +207,67 @@ class ApiCommandBase extends CommandBase {
       $param = $input->getOption($param_name);
     }
     return $param;
+  }
+
+  /**
+   * @param \Symfony\Component\Console\Input\InputInterface $input
+   * @param \Symfony\Component\Console\Output\OutputInterface $output
+   *
+   * @throws \Exception
+   */
+  protected function fillMissingApplicationUuid(InputInterface $input, OutputInterface $output): void {
+    if ($input->hasArgument('applicationUuid') && !$input->getArgument('applicationUuid')) {
+      $output->writeln('Inferring Cloud Application UUID for this command since none was provided...');
+      if ($application_uuid = $this->determineCloudApplication()) {
+        $output->writeln("Set application uuid to <options=bold>$application_uuid</>");
+        $input->setArgument('applicationUuid', $application_uuid);
+      }
+    }
+  }
+
+  /**
+   * @param \Symfony\Component\Console\Input\InputInterface $input
+   *
+   * @throws \Acquia\Cli\Exception\AcquiaCliException
+   */
+  protected function convertApplicationAliastoUuid(InputInterface $input) {
+    if ($input->hasArgument('applicationUuid') && $input->getArgument('applicationUuid')) {
+      $application_uuid_argument = $input->getArgument('applicationUuid');
+      try {
+        $this->validateUuid($application_uuid_argument);
+      } catch (ValidatorException $exception) {
+        // Since this isn't a valid UUID, let's see if it's a valid alias.
+        try {
+          $customer_application = $this->getApplicationFromAlias($application_uuid_argument);
+          $input->setArgument('applicationUuid', $customer_application->uuid);
+        } catch (ValidatorException $exception) {
+          // Let Cloud API handle throwing an exception from here.
+        }
+      }
+    }
+  }
+
+  /**
+   * @param \Symfony\Component\Console\Input\InputInterface $input
+   *
+   * @throws \Acquia\Cli\Exception\AcquiaCliException
+   */
+  protected function convertEnvironmentAliasToUuid(InputInterface $input): void {
+    if ($input->hasArgument('environmentId') && $input->getArgument('environmentId')) {
+      $env_uuid_argument = $input->getArgument('environmentId');
+      try {
+        $this->validateUuid($env_uuid_argument);
+      } catch (ValidatorException $exception) {
+        // Since this isn't a valid UUID, let's see if it's a valid alias.
+        try {
+          $this->validateEnvironmentAlias($env_uuid_argument);
+          $environment = $this->getEnvironmentFromAliasArg($env_uuid_argument);
+          $input->setArgument('environmentId', $environment->uuid);
+        } catch (ValidatorException $exception) {
+          // Let Cloud API handle throwing an exception from here.
+        }
+      }
+    }
   }
 
 }
